aviation_assistant/storage/document_store.py
from typing import Dict, Any, List, Optional
import logging
import os
import pickle
import numpy as np
from datetime import datetime

import faiss
from haystack.document_stores import FAISSDocumentStore
from haystack.schema import Document

class AviationDocumentStore:
    """
    Document store for aviation data using FAISS for vector storage.
    
    Stores and retrieves aviation documents including weather, NOTAMs,
    and airport/route information.
    """
    
    def __init__(self, 
                embedding_dim: int = 384, 
                faiss_index_factory_str: str = "Flat",
                vector_db_path: Optional[str] = None):
        """
        Initialize the aviation document store.
        
        Args:
            embedding_dim: Dimension of the embeddings
            faiss_index_factory_str: FAISS index type
            vector_db_path: Optional path to save the vector database
        """
        self.logger = logging.getLogger(__name__)
        self.vector_db_path = vector_db_path
        
        # Initialize the FAISS document store
        self.document_store = FAISSDocumentStore(
            embedding_dim=embedding_dim,
            faiss_index_factory_str=faiss_index_factory_str,
            return_embedding=True,
            sql_url="sqlite:///faiss_document_store.db" if vector_db_path else None
        )
    
    def add_documents(self, documents: List[Dict[str, Any]], embeddings: List[List[float]]) -> None:
        """
        Add documents with embeddings to the document store.
        
        Args:
            documents: List of document dictionaries
            embeddings: List of document embeddings
        """
        # TODO: Implement this method
        pass
    
    def query_documents(self, 
                       query_embedding: List[float], 
                       top_k: int = 5,
                       filters: Optional[Dict[str, Any]] = None) -> List[Dict[str, Any]]:
        """
        Query documents using embedding similarity.
        
        Args:
            query_embedding: Query embedding vector
            top_k: Number of results to return
            filters: Optional filters to apply
            
        Returns:
            List of matching documents
        """
        # TODO: Implement this method
        pass
    
    def save(self) -> None:
        """Save the document store to disk."""
        if self.vector_db_path:
            self.document_store.save(self.vector_db_path)
            self.logger.info(f"Saved document store to {self.vector_db_path}")
    
    def load(self) -> None:
        """Load the document store from disk."""
        if self.vector_db_path and os.path.exists(self.vector_db_path):
            self.document_store.load(self.vector_db_path)
            self.logger.info(f"Loaded document store from {self.vector_db_path}")
