requirements.txt
# Core
python-dotenv==1.0.0
pydantic==2.4.2

# Data Collection
requests==2.31.0
aiohttp==3.8.5

# Data Processing & RAG
haystack==2.0.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==1.3.5

# Visualization
matplotlib==3.8.0
plotly==5.18.0
folium==0.14.0

# UI
streamlit==1.27.2

# Utilities
tqdm==4.66.1
loguru==0.7.2
